feat : 간단한 주문 -> 결제 비즈니스 로직 추가 및 빈 충돌 방지로 idempotency 접미사를 가진 빈이 생성되도록 함

+++ src/main/java/com/taehyun/db_playground/redis/lock/scenario01/domain/Wallet.java
@@ -30,7 +30,7 @@ public class Wallet {
         return new Wallet(userId, balance);
     }
 
-    public void deduct(int amount) {
+    public void deduct(long amount) {
 
         if (amount == 0)
             throw new IllegalArgumentException("차감할 금액이 0 일 수 없습니다");
